브레즌햄 알고리즘 수정

+++ GameProject/06_JumpPointSearch/project/project/CBresenham.h
@@ -23,7 +23,7 @@ public:
 
 	int addNum;
 	int error;
-	int errorMax;
+	float errorMax;
 
 	int deltaX;
 	int deltaY;
